# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Purple Squad API
  version: v1
  description: API for the Purple Squad home services platform.
paths:
...
  /api/v1/admin/payments/{id}/reconcile-refund/:
    post:
      operationId: api_v1_admin_payments_reconcile_refund_create
      summary: Reconcile a refund with Razorpay
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: ''
  /api/v1/admin/payments/{id}/refund/:
    post:
      operationId: api_v1_admin_payments_refund_create
      summary: Refund a successful Razorpay payment
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RefundCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RefundCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: ''
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        booking:
          type: string
          format: uuid
          readOnly: true
        booking_number:
          type: string
          readOnly: true
        customer_phone:
          type: string
          readOnly: true
        provider:
          allOf:
          - $ref: '#/components/schemas/ProviderEnum'
          readOnly: true
        provider_order_id:
          type: string
          readOnly: true
          nullable: true
        provider_payment_id:
          type: string
          readOnly: true
          nullable: true
        provider_refund_id:
          type: string
          readOnly: true
          nullable: true
        parent_payment:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        currency:
          type: string
          readOnly: true
        payment_type:
          allOf:
          - $ref: '#/components/schemas/PaymentTypeEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/PaymentRecordStatusEnum'
          readOnly: true
        signature_verified:
          type: boolean
          readOnly: true
        idempotency_key:
          type: string
          readOnly: true
        paid_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        refunded_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - booking
      - booking_number
      - created_at
      - currency
      - customer_phone
      - id
      - idempotency_key
      - paid_at
      - parent_payment
      - payment_type
      - provider
      - provider_order_id
      - provider_payment_id
      - provider_refund_id
      - refunded_at
      - signature_verified
      - status
      - updated_at
...
    PaymentRecordStatusEnum:
      enum:
      - CREATED
      - PENDING
      - SUCCESS
      - FAILED
      - REFUNDED
      type: string
      description: |-
        * `CREATED` - Created
        * `PENDING` - Pending
        * `SUCCESS` - Success
        * `FAILED` - Failed
        * `REFUNDED` - Refunded
...
    PaymentTypeEnum:
      enum:
      - BOOKING_ADVANCE
      - BALANCE
      - REFUND
      type: string
      description: |-
        * `BOOKING_ADVANCE` - Booking advance
        * `BALANCE` - Balance
        * `REFUND` - Refund
...
    ProviderEnum:
      enum:
      - RAZORPAY
      - OFFLINE
      type: string
      description: |-
        * `RAZORPAY` - Razorpay
        * `OFFLINE` - Offline
    RefundCreateRequest:
      type: object
      properties:
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        reason:
          type: string
          maxLength: 500
      required:
      - amount
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT